added correct naviation routes
